Add support for objects systematics
For each systematics, CMSSW modules are automatically added to the
sequence to produce the systematically modified objects (shifted jets
for example). For each newly added, a Framework producer is added, named
like the nominal producer but with a postfix ('jets' -> 'jets_jecup').

Analyzers configurations using any of the nominal producers are cloned,
and parameters are automatically updated to match the new producers. In
order for this to work, your analyzer must have some parameters of type
`string` or `InputTag` whose value the name of the producer.

Right now, only JEC (JES) and JER are supported, but the system is
generic enough to add more if they are one day needed.